LPS-63528 Move Constants to model package
